ported rest of types
