+++ src/samples/DistanceImage.hpp
@@ -5,9 +5,7 @@
 #include <base/Eigen.hpp>
 #include "Pointcloud.hpp"
 #include <vector>
-#include <Eigen/Core>
 #include <boost/math/special_functions/fpclassify.hpp>
-#include <limits>
 
 namespace base
 {
@@ -67,10 +65,7 @@ namespace samples
 	/** 
 	 * resets all values in the distance image to nan
 	 */
-	void clear()
-	{
-	    std::fill( data.begin(), data.end(), std::numeric_limits<float>::quiet_NaN() );
-	}
+	void clear();
 
 	/** 
 	 * Transforms a x, y pixel coordinates into a 3d scene point, using the distance
@@ -122,19 +117,7 @@ namespace samples
 	    return true;
 	}
 
-        base::samples::Pointcloud getPointCloud() const
-        {
-            base::samples::Pointcloud pointCloud;
-            Eigen::Vector3d point;
-            for(size_t y = 0; y < height ; y++)
-            {
-                for(size_t x = 0; x < width ; x++)
-                {
-                    pointCloud.points.push_back(point);
-                }
-            }
-            return pointCloud;
-        }
+        base::samples::Pointcloud getPointCloud() const;
         
 	/** 
 	 * The intrinsic matrix has the following form
@@ -177,13 +160,7 @@ namespace samples
 	 * @param c_x center point in x
 	 * @param c_y center point in y
 	 */
-	void setIntrinsic( double f_x, double f_y, double c_x, double c_y )
-	{
-	    scale_x = 1.0 / f_x;
-	    scale_y = 1.0 / f_y;
-	    center_x = -c_x / f_x;
-	    center_y = -c_y / f_y;
-	}
+	void setIntrinsic( double f_x, double f_y, double c_x, double c_y );
 
 	/** @brief set the size of the distance image
 	 *
@@ -193,12 +170,7 @@ namespace samples
 	 * @param width width of the image
 	 * @param height height of the image
 	 */
-	void setSize( double width, double height )
-	{
-	    this->width = width;
-	    this->height = height;
-	    data.resize( width * height );
-	}
+	void setSize( double width, double height );
     };
 }
 }
